% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_best_marginal_model.R
\name{build_marginal_model}
\alias{build_marginal_model}
\title{Build marginal model}
\usage{
build_marginal_model(load_name, f_name, mle = NULL, wf_mom = 0,
  wf_feat = 0, max_iter = 50, transformation = NULL,
  pick_only_transformation = NULL, general_path, eqs_df = NULL,
  starting_model = NULL)
}
\arguments{
\item{f_name}{name of the file where to save spike count optimal model}

\item{mle}{whether is mle or kf}

\item{wf_mom}{waveform moment, if 0 computes for spike counts}

\item{wf_feat}{waveform feature, if 0 computes for spike counts}

\item{max_iter}{number of max iteration before stopping}

\item{transformation}{transformation selected 1 linear, 2 ace, 3 sqrt, NULL all}

\item{pick_only_transformation}{in case you only want to select a subset of
the data, indicate which transformation in particular}

\item{general_path}{path for saving the file}

\item{load_file}{file to all with all data}
}
\description{
This function builds the marginal model for spike count model and all
waveoforms
}
