% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_plus_prune.R
\name{combine_and_prune}
\alias{combine_and_prune}
\title{Combine and prune models}
\usage{
combine_and_prune(load_name, sc_fname, wf_fname, save_name_partial,
  save_name_final, mle = NULL, general_path,
  pick_only_transformation = NULL)
}
\arguments{
\item{load_name}{file to all with all data}

\item{sc_fname}{name of the file where to save spike count optimal model}

\item{wf_fname}{name of the file where to save waveform optimal model}

\item{save_name_partial}{name of the name where to store partial results}

\item{save_name_final}{name of the name where to store final results}

\item{mle}{whether is mle or kf}

\item{general_path}{what's the path for sc_fname and wf_fname}
}
\description{
This function combines and prunes the marginal models
}
