% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_functions.R
\name{do_backward}
\alias{do_backward}
\title{Backward mse}
\usage{
do_backward(new_reference, ztrain, ztest, A, W, H, Q, model, mle, int, Time,
  vel, condition)
}
\arguments{
\item{new_reference}{error of the current model}

\item{ztrain}{training neural data}

\item{ztest}{testing neural data}

\item{A}{the state matrix}

\item{W}{the state covariance matrix}

\item{H}{the encoding matrix}

\item{Q}{the encoding covariance matrix}

\item{model}{given model}

\item{mle}{whether mle or kf}

\item{int}{vector of integer on when the trial starts}

\item{Time}{vector of time of the intervals}

\item{vel}{is the reference velocity to use, need to be compatible with int
and Time}

\item{condition}{whether to prune only on some equation}
}
\value{
a list with pruned model and current risk
}
\description{
This function computes leave one out mse for a given model
}
