% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model_functions.R
\name{get_best_model}
\alias{get_best_model}
\title{Computing the best sc model for each electrode}
\usage{
get_best_model(ztrain, ztest, model, n_units, n_transf = 3, n_lags = 13,
  eqs_df, max_iter = 10, mle = FALSE, save_name, random_order = TRUE,
  int_dec, time_dec, vel, veltest, is_sc = TRUE, wf_mom = NULL,
  wf_feat = NULL, transf_selected = NULL)
}
\arguments{
\item{ztrain}{neural data in the training set}

\item{ztest}{neural data in the testing set}

\item{model}{starting model, one per electrode}

\item{n_units}{number of units in the dataset}

\item{n_transf}{number of sc transforamtion used, by default 3}

\item{n_lags}{number of lags used, by default 13}

\item{eqs_df}{data frame for equation reference}

\item{max_iter}{max iteration of the algorithm if it doesn't converge first}

\item{mle}{whether mle is true or false}

\item{save_name}{file name where saving the objects}

\item{random_order}{whether to iterate randomly over the electrodes}

\item{int_dec}{vector of integer on when the trial starts}

\item{time_dec}{vector of time of the intervals}

\item{vel}{velocity in the training set}

\item{veltest}{velocity in the testing set}

\item{is_sc}{whether the best model should run on spike counts, default equal TRUE}

\item{wf_mom}{wf moment to consider, only if is_sc is FALSE}

\item{transf_selected}{transformation selected 1 linear, 2 ace, 3 sqrt, NULL all}

\item{wf_key}{wf key to consider, only if is_sc is FALSE}
}
\value{
the final model
}
\description{
This function computes the best sc model for each electrode, following three
steps:
1. start with the initial model at the best uniform lag;
2. for each electrode in turn, evaluate the risks of the 3*13 models
that correspond to using the 3*13 SC equations for that electrode, and replace
	the current equation with the best equation.
}
