% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_functions.R
\name{get_kf_score}
\alias{get_kf_score}
\title{Computes the kf score}
\usage{
get_kf_score(zcenter, A, W, H, Q, Qm1, current_size, full_size, velcenter, int,
  Time, method)
}
\arguments{
\item{zcenter}{centered neural data}

\item{A}{the state matrix}

\item{W}{the state covariance matrix}

\item{H}{the encoding matrix}

\item{Q}{the encoding covariance matrix}

\item{Qm1}{inverse of the encoding covariance matrix}

\item{current_size}{size of the current model}

\item{full_size}{size of the full model}

\item{int}{vector of integer on when the trial starts}

\item{Time}{vector of time of the intervals}

\item{method}{forward for add one in and backward for leave one out}

\item{velcener}{velocity in the training set}
}
\value{
the average kf score for each trial left-out or add-in
}
\description{
This function computes the kf score, based on the method chosen
}
