% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_functions.R
\name{get_mle_loo_score}
\alias{get_mle_loo_score}
\title{Computes the mle leave-one-out mean error for each equation left out}
\usage{
get_mle_loo_score(zcenter, H, Q, Qm1, current_size, full_size, velcenter, int,
  Time)
}
\arguments{
\item{zcenter}{centered neural data}

\item{H}{the encoding matrix}

\item{Q}{the encoding covariance matrix}

\item{Qm1}{inverse of the encoding covariance matrix}

\item{current_size}{size of the current model}

\item{full_size}{size of the full model}

\item{velcenter}{centered  velocity in the training set,
if zcenter is from the training set}

\item{int}{vector of integer on when the trial starts}

\item{Time}{vector of time of the intervals}

\item{mle}{boolean TRUE if mle, FALSE if KF}
}
\value{
a vector of mean error for each equation left out
}
\description{
This function computes mle leave-one-out mean error for each equation left out
}
