% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_functions.R
\name{get_score}
\alias{get_score}
\title{Leave-one-out/Add-one-in score wrapper function}
\usage{
get_score(ztrain, ztest = NULL, A = NULL, W = NULL, H = NULL,
  Q = NULL, Qm1 = NULL, decoding_train = FALSE, int_train = int.Train,
  time_train = Time.Train, int_test = int.Test, time_test = Time.Test,
  vel = velocity.Train, veltest = velocity.Test, full_size = NULL,
  current_size = NULL, mle, method = "forward")
}
\arguments{
\item{ztrain}{neural data in the training set}

\item{ztest}{neural data in the testing set}

\item{A}{the state matrix}

\item{W}{the state covariance matrix}

\item{H}{the encoding matrix}

\item{Q}{the encoding covariance matrix}

\item{Qm1}{inverse of the encoding covariance matrix}

\item{int_train}{vector of integer on when the trial starts in the training set}

\item{time_train}{vector of time of the intervals in the training set}

\item{int_test}{vector of integer on when the trial starts in the testing set}

\item{time_test}{vector of time of the intervals in the testing set}

\item{vel}{velocity in the training set}

\item{full_size}{size of the full model}

\item{current_size}{size of the current model}

\item{mle}{boolean TRUE if mle, FALSE if KF}

\item{method}{forward for add one in and backward for leave one out}
}
\value{
the average mse score for each trial left-out or add-in
}
\description{
This functions wrapper function that computes leave-one-out/add-one-in score
in either MLE or KF settings calling the right functions depending on the settings
}
