% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_functions.R
\name{prune_model}
\alias{prune_model}
\title{Prune the model}
\usage{
prune_model(ztrain, ztest, model, int, Time, vel, mle, condition = 0)
}
\arguments{
\item{ztrain}{training neural data}

\item{ztest}{testing neural data}

\item{int}{vector of integer on when the trial starts}

\item{Time}{vector of time of the intervals}

\item{vel}{is the reference velocity to use, need to be compatible with int
and Time}

\item{mle}{whether mle or kf}

\item{condition}{whether to prune only on some equation}
}
\value{
a list with pruned model and current risk
}
\description{
This function prunes the current model
}
