% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_functions.R
\name{show_condition}
\alias{show_condition}
\title{This functions checks whether a code produces error/warning/message}
\usage{
show_condition(code)
}
\arguments{
\item{code}{code to test}
}
\value{
string with either error/warning/message
}
\description{
This functions checks whether a code produces error/warning/message
}
