% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model_functions.R
\name{update_model}
\alias{update_model}
\title{Updates the model if risk is lowered}
\usage{
update_model(model, eqs_df, unit, wwhich, candidate_eqs, mle, ztrain, ztest,
  Qm1)
}
\arguments{
\item{model}{current model}

\item{eqs_df}{data frame with equation info}

\item{unit}{unit selected}

\item{wwhich}{which are the candidate to replace that given equation}

\item{candidate_eqs}{list of all candidate eqs}

\item{mle}{whether mle or kf}

\item{ztrain}{neural data in the training set}

\item{ztest}{neural data in the testing set}

\item{Qm1}{inverse of the covariance matrix}
}
\value{
a list of vectors update and needed for next iteration
}
\description{
This function udates the model if the risk is lowered by one of the 13*3 equations
that belong to the same electrode
}
